+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ### Features
 - Added custom tooltip GUI scaling option (Just really missed that)
+- Added full inventory alert option
 
 ### Fixes
 - Made item models and custom tooltip styles fall back to vanilla defaults if